Updated docker-compose.yml, fixed open-questions new writes to db bug

+++ postgresql/init.sql
@@ -1,5 +1,5 @@
-CREATE TABLE questions(id SERIAL PRIMARY KEY, userid INT, header VARCHAR, question VARCHAR, status BOOLEAN, solverid INT, solution VARCHAR);
+CREATE TABLE questions(id SERIAL PRIMARY KEY, userid BIGINT, header VARCHAR, question VARCHAR, status BOOLEAN, solverid BIGINT, solution VARCHAR);
 
-CREATE TABLE solutions(id SERIAL PRIMARY KEY, questionid INT, solverid INT, solution VARCHAR);
+CREATE TABLE solutions(id SERIAL PRIMARY KEY, questionid INT, solverid BIGINT, solution VARCHAR);
 
-CREATE TABLE banned(id SERIAL PRIMARY KEY, userid INT, reason VARCHAR);
+CREATE TABLE banned(id SERIAL PRIMARY KEY, userid BIGINT, reason VARCHAR);
